Modify snapshot_context structure for persistency
Previous snapshot_context interface is insufficient for decoupling
homestore from customers on snapshot implementations. This commit
replaces deserialize() with a virtual function of ReplDev for
constructing particular snapshot_context instance from byte buffer.
