+++ src/lib/replication/repl_dev/raft_repl_dev.h
@@ -113,6 +113,38 @@ struct ReplDevCPContext {
     uint64_t last_applied_dsn;
 };
 
+class nuraft_snapshot_context : public snapshot_context {
+public:
+    nuraft_snapshot_context(nuraft::snapshot &snp) : snapshot_context(snp.get_last_log_idx()) {
+        auto snp_buf = snp.serialize();
+        snapshot_ = nuraft::snapshot::deserialize(*snp_buf);
+    }
+
+    nuraft_snapshot_context(sisl::io_blob_safe const &snp_ctx) : snapshot_context(0) { deserialize(snp_ctx); }
+
+    sisl::io_blob_safe serialize() override {
+        // Dump the context from nuraft buffer to the io blob.
+        auto snp_buf = snapshot_->serialize();
+        sisl::io_blob_safe blob{s_cast<size_t>(snp_buf->size())};
+        std::memcpy(blob.bytes(), snp_buf->data_begin(), snp_buf->size());
+        return blob;
+    }
+
+    void deserialize(const sisl::io_blob_safe &snp_ctx) {
+        // Load the context from the io blob to nuraft buffer.
+        auto snp_buf = nuraft::buffer::alloc(snp_ctx.size());
+        snp_buf->put_raw(snp_ctx.cbytes(), snp_ctx.size());
+        snp_buf->pos(0);
+        snapshot_ = nuraft::snapshot::deserialize(*snp_buf);
+        lsn_ = snapshot_->get_last_log_idx();
+    }
+
+    nuraft::ptr<nuraft::snapshot> nuraft_snapshot() { return snapshot_; }
+
+private:
+    nuraft::ptr<nuraft::snapshot> snapshot_;
+};
+
 class RaftReplDev : public ReplDev,
                     public nuraft_mesg::mesg_state_mgr,
                     public std::enable_shared_from_this< RaftReplDev > {
@@ -204,8 +236,13 @@ public:
         m_data_journal->purge_all_logs();
     }
 
+    std::shared_ptr<snapshot_context> deserialize_snapshot_context(sisl::io_blob_safe &snp_ctx) override {
+        return std::make_shared<nuraft_snapshot_context>(snp_ctx);
+    }
+
     //////////////// Accessor/shortcut methods ///////////////////////
-    nuraft_mesg::repl_service_ctx* group_msg_service();
+    nuraft_mesg::repl_service_ctx *group_msg_service();
+
     nuraft::raft_server* raft_server();
     RaftReplDevMetrics& metrics() { return m_metrics; }
